feat: ActivityPub service (#66)
Implemented an ActivityPub service that encapsulates an inbox and outbox. The service allow clients to post messages to the outbox and register for notifications of inbound activities. The service provides an activity handler to process the activities. Currently only a skeleton implementation is provided for the 'Create' activity. Future commits will provide implementations for all activities.

closes #56

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>