feat: Handler for 'Follow' activity
Implemented a handler for the ActivityPub 'Follow' activity. An authorization handler is 'asked' if the service should accept or reject the follow request. If accepted, the handler posts an 'Accept' activity to the requesting actor and adds the actor to the list of followers; otherwise a 'Reject' activity is posted.

closes #69

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>