feat: Handler for the ActivityPub 'Offer' activity
Implemented a handler for the ActivityPub 'Offer' activity. The handler invokes a 'witness' handler with the embedded anchor credential so that it may generate a proof. If successful, the handler posts a 'Like' activity with the proofs back to the actor (requestor). The activity also adds the activity to the service's 'liked' list.

closes #72

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>