fix: Offer handler responds with incorrect start and end time in Like activity
The Offer handler responds with a 'Like' activity and the start and end times were set incorrectly. This commit changes the end time to be the current server time plus a configurable delay.

Also, the Offer activity is rejected if the offer has already expired (according to the endTime in the Offer activity).

closes #81

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>