feat: '/followers' REST endpoint
Added a REST endpoint that returns a service's followers as a Collection or CollectionPage type. If the request does not include a 'page' parameter then a 'Collection' is returned that specifies the first and last page of the followers. If the request contains a 'page=true' parameter then a 'CollectionPage' is returned containing the IRI's of the followers for that page, along with the previous and next page.

closes #95

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>