feat: '/following' REST endpoint
Added a REST endpoint that returns the URIs that a service is following as a Collection or CollectionPage type. If the request does not include a 'page' parameter then a 'Collection' is returned that specifies the first and last page of the following list. If the request contains a 'page=true' parameter then a 'CollectionPage' is returned containing the IRI's of the list of 'following' for that page, along with the URLs of the previous and next page.

closes #96

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>