feat: Implemented '/outbox' and '/inbox' REST endpoints
Added REST endpoints that return the activities in the service's inbox and outbox as an OrderedCollection or OrderedCollectionPage type. If the request does not include a 'page' parameter then an 'OrderedCollection' is returned that specifies the first and last page of the inbox or outbox. The first page contains the most recent activities. If the request contains a 'page=true' parameter then an 'OrderedCollectionPage' is returned containing the activities for that page, along with the URLs of the previous and next page.

closes #93
closes #94

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>