feat: Implemented '/witnesses' and '/witnessing' REST endpoints
Added REST endpoints that return a service's witnesses as well as the URIs that the local service is witnessing. If the request does not include a 'page' parameter then a 'Collection' is returned that specifies the first and last page of the witnesses/witnessing. If the request contains a 'page=true' parameter then a 'CollectionPage' is returned containing the witnesses/witnessing URIs for that page, along with the URIs of the previous and next page.

closes #97
closes #98

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>