feat:  '/liked' REST endpoint
Added a REST endpoint that returns the URIs of the objects that the service 'liked' as an OrderedCollection or OrderedCollectionPage type. If the request does not include a 'page' parameter then an 'OrderedCollection' is returned that specifies the first and last page of the objects that were liked. The most recent 'liked' objects are in the first page. If the request contains a 'page=true' parameter then an 'OrderedCollectionPage' is returned containing the IRI's of the liked for that page, along with the previous and next page.

closes #100

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>