feat: Outbox Post sets the ID and 'actor' of activity
The Outbox.Post signature has changed such that the ID of the activity is returned upon a successful post. When posting an activity to the Outbox then, if the 'id' field hasn't already been set, the Outbox sets the 'id' field of the activity to a unique ID. Also, if the 'actor' field is set then the Outbox validates that it is set to the service IRI otherwise an error is returned. If the 'actor' field is not set then it is set to the service IRI.

closes #182

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>