fix: Publish Announce activity to followers
The 'to' field of the 'Announce' activity is now set to the actor's '/followers' collection as opposed to setting it to the IRIs of each individual follower. This may result in an 'Announce' activity being posted back to the originator of the 'Create', but there is already handling for duplicate anchor credential at the originator, so this will not be a problem.

closes #347

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>