fix: Anchor Origin MUST be provided for Create and Recover
Currently anchor origin is checked against allowed list  if it is provided. However if it is not (nil) it will go through. During early drafts of Orb spec there was an idea about filling in anchor origin if not provided. This requirement got dropped (evolved) later on and anchor origin validator never got updated.

Closes #403

Signed-off-by: Sandra Vrtikapa <sandra.vrtikapa@securekey.com>