feat: Local CAS now uses IPFS-compatible CIDs
The local CAS now uses CIDs that are compatible with IPFS, assuming that the IPFS node is running under default settings and that the data is less than 256KB.

A new flag was added to choose between v1 or v0 CIDs. The IPFS client will also use this format now. v1 CIDs are the default.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>