feat: Add bearer token authorization for /cas endpoint
If a bearer token is defined for the /cas endpoint then the client must provide a valid token in the header. If a token is not found in the header then the HTTP signature in the header is verified and the actor of the signature is verified to be either a follower or witness.

closes #408

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>