chore: Move updating did anchor references to observer
Move updating anchor references from writer to observer. This will allow for haveing up-to-date did anchor references for following domains too so we will be able to send did updates to following domains.

Add test for updating did document in different domain from where it was created. Of course that domain has to follow original domain.

Closes #441

Signed-off-by: Sandra Vrtikapa <sandra.vrtikapa@securekey.com>