chore: Ensure transient errors are retried by AMQP message queue
Transient errors are returned for operations that could succeed with a retry. A test case was added to ensure that messages which are 'Nacked' are retried by the AMQP message broker.

Also added configuration for RabbitMQ in the BDD test, as well as changed the image to use one which includes the management plugin.

closes #472

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>