fix: Don't check if actor in offer activity is being witnessed
In the offer activity handler, remove the check that ensures the actor is in the 'witnessing' collection since a witness can be explicitly added to an operation (without being a system witness).

Actor verification for the /cas endpoint is turned off since now the actor is not in the 'witnessing' collection but needs access to the /cas endpoint.

closes #525

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>