fix: Add Nack to operation queue
Upgraded to the latest sidetree-core-go which defines an explicit 'Nack' for the operation queue when operations are removed. If an error occurs then Nack will be invoked and the removed AMQP messages will also be nacked and the messages will be retried on another subscriber. The message broker may be configured to limit the number of retries.

closes #524

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>