test: Re-enable IPNS tests and skip upload if already there
IPNS upload is now a step at the beginning of the DID tests. This makes it so the IPNS document only gets uploaded for the specific tests that it's needed for. IPNS uploading will also be skipped if the document can already be found at the time of the test.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>