chore: Additional logs for proof handler
Added Info logs and changed existing Debug logs to Info in order to help diagnose a race condition that may occur when multiple proofs for a VC are sent to different server instances in the same domain. The result of this race is that both instances think that additional proofs are required and the VC is never anchored.

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>