fix: Handle null return value from IPFS
In certain cases null is returned from IPFS when the CID is not found. (In other cases the request times out.) When null is returned then return a "content not found" error to the caller.

Also updated the 'create' activity in the BDD test to embed an anchor credential since the test would sometimes time out on IPFS since the anchor credential was never stored to IPFS.

closes #727

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>