fix: Some expiry service unit tests failing intermittently
Some unit tests for the expiry service code relied on starting the service, waiting a short period of time, and then checking the logs to see if an expected failure message was logged. If a system is running slowly, these failure messages may not get logged in time, resulting in a failure. Now, the test will check the logs a maximum of 20 times (with a delay between each check) before giving up and concluding the test has failed. This allows the tests to pass on slower systems while still running quickly on faster machines.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>