fix: Intermittent BDD test failure
The issue was that orb2-domain1 didn't have the REPLICATE_LOCAL_CAS_WRITES_IN_IPFS setting enabled - it was only enabled on orb-domain1. In one of the tests, the create operation would happen on orb2-domain1 (which didn't have IPFS replication on) and then the update would happen on orb-domain1 (which did). Since the create operation wasn't replicated in IPFS, the IPFS DID discovery failed.

There needs to be further investigation into why orb2-domain1 was getting the create request since in the BDD test it appears that it should have been sent to orb.domain1. We also need to figure out why it would sometimes pass.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>