feat: Use an AMQP channel pool for the publisher
Use the latest watermill-amqp client which implements a Publisher channel pool. The channel pool improves performance in high volume systems by not incurring the overhead of opening and closing channels, and the pool also acts as a throttle to prevent the AMQP server from overloading.

Also, use separate AMQP connections for publishers and subscribers since this is recommended for RabbitMQ (https://www.cloudamqp.com/blog/part1-rabbitmq-best-practice.html#separate-connections-for-publisher-and-consumer).

closes #943
closes #985

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>