chore: Add 0 or positive check for minimum number of witnesses required.
Currently for minimum number of witnesses in witness policy rule OutOf() we only check for integer for first parameter. It should be 0 or positive integer since it denotes minimum number of witnesses required to satisfy policy.

Percent rule MinPercent() is already enforcing integer that is between 0 and 100 which is correct for percent.

Closes #1025

Signed-off-by: Sandra Vrtikapa <sandra.vrtikapa@securekey.com>