fix: Recovery of anchors from anchor origin
This commit adds the capability for an anchor origin to recover AnchorEvents and Sidetree files that it had previously published but does not currently have in its database (possibly due to a backup/restore). Logic has been added in the anchor sync process to also process the inboxes of services that are following me. So, an AnchorEvent may be recovered by reading the inbox of a service to which I had previously published a 'Create' activity. All of the Sidetree files are also recovered by supplying an 'alternate source' to the Observer. (The alternate source is the service from which I recovered the 'Create' activity.)

Also upgraded the Go version to 1.17 to all Docker images/builds and updated the Alpine version to 3.15. (This fixes a bug in Go version 1.17.3 where closing an HTTP response body always resulted in an error 'context closed'. See https://github.com/golang/net/commit/f0573a1506aed9f8168fbd7d5e757448db58b427.)

closes #1030

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>