fix: Add minimum activity age param to anchor sync task
A minimum activity age parameter is added to the anchor sync task such that the activity needs to be older than this value before it is processed (according to the "published" time of the activity).

For example, the anchor sync task should not process activities that are younger than 1 minute. The minimum age is configurable via the ANCHOR_EVENT_MIN_ACTIVITY_AGE startup parameter (default 1m).

closes #1085

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>