feat: Check if anchor exists if 0 operations processed
If the transaction processor determines that all of the operations in the batch are already in local storage then a check is made to see if the anchor is also store. If so, the anchor will not be processed again. (This is most likely due do multiple Announce activities announcing the same anchor event.)

closes #1104

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>