fix: Unit test MongoDB and RabbitMQ port allocation errors
For some reason, when running our unit tests the MongoDB and RabbitMQ container port allocations may not be released (or may take some time to be released), resulting in errors for later tests. The code that starts up the MongoDB Docker image tries different ports until it finds a free one, but the existing attempt limit of 3 seems to be insufficient now, so I bumped it up to 10. I also added the same type of port-retry behaviour for the tests that start up RabbitMQ Docker containers, as they suffer from the same problem.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>