test: Updated onboarding and recover BDD test to handle instance failure
Modified the existing onboarding-and-recovery BDD test to take down an Orb instance while DIDs were being created in order to prove that no messages are lost in a busy system when a domain instance goes down.

Also added a fix to the HTTP signature verifier to propagate a transient HTTP error to the caller so that a transient error may be retried.

closes #1225

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>