fix: Start Orb services after the REST endpoints
The REST endpoints must be up before the background services are started since a background service may trigger an action on an external domain that causes that domain to call back to a local REST endpoint (which may not be available yet).

Also, removed debug logs for Log monitor since they weren't providing any relevant information.

closes #1253

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>