feat: Enable AMQP publisher delivery confirmation
When an AMQP server is restarted, the Orb pub-sub client receives a success response when publishing a message even though the message may not have been persisted by the AMQP server (because the server has died) and therefore the message is lost. This commit uses the latest version of watermill-amqp which adds a delivery confirmation feature. A failed delivery confirmation results in the message being retried (according to the pub-sub redelivery mechanism).

A new startup parameter was added to enable/disable delivery confirmations. Delivery confirmations are enabled by default.

(Note that the delivery confirmation pull request (https://github.com/ThreeDotsLabs/watermill-amqp/pull/15) has not yet been merged, so this commit points to the TrustBloc fork of watermill-amqp. Once the PR is merged, Orb should be updated to point to the official library.)

Also:
- Added a BDD test for AMQP server restart
- Changed the onboarding-and-recovery BDD test to be run only for local CAS in order to reduce the time to run the integration tests

closes #1290

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>