fix: Panic in HTTP subscriber Stop
Added a publisher channel to handle publish requests in a separate Go proc. When the HTTP subscriber is stopped, the publisher listener safely shuts down the subscriber channel.

This commit also fixes the problem with varying code coverage in the unit tests and a race condition in the startup params test.

closes #1463
closes #591

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>