fix: Make VC issuer the same as author of anchor linkset
The issuer in the anchor credential (VC) is now always the same as the author of the anchor linkset. So if the auther is a DID then the issuer will be set to the same DID.

This commit also removes the startup parameters, anchor-credential-issuer and anchor-credential-url since they are redundant and can also lead to misconfiguration.

closes #1488

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>