fix: Ensure parent anchor is processed only once
Added a "status" field to the anchor-ref store such that the status is set to PENDING before the anchor is posted to the anchor queue for processing. This status is checked before the parent anchor is posted to the queue for processing. This ensures that other threads or instances do not post the same parent to the queue multiple times.

closes #1562

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>