fix: Ensure update operations are processed after create operations are published
Publish an "Update" operation to the queue with a delay if a "Create" operation with the same suffix has not yet been published.

Also de-fragment the operation queue such that operations that have previously failed are grouped
contiguously and operations that have no error (thus far) are also grouped contiguously. Grouping (de-fragmenting)
the operation queue provides a greater chance that valid operations are successfully processed and are not
failed just because the batch failed due to a single "bad" operation.

closes #1583

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>